first implementation of RC24 dark (#42)
* first implementation of RC24 dark

May or may not break things, also may or may not be incomplete. Will not be merged into master for RiiConnect24 until complete.

* first implementation of RC24 dark (upload 2)

May or may not break things, also may or may not be incomplete. Will not be merged into master for RiiConnect24 until complete.

* I uploaded it to the wrong place woops

* first implementation of RC24 dark (upload 3)

May or may not break things, also may or may not be incomplete. Will not be merged into master for RiiConnect24 until complete.

* Cleaned up some of the CSS

* More cleaning - Not removing any of the !important

* yet more cleaning - still not removing the !important